installkernel-bls: remove unnecessary check for GRUB_ENABLE_BLSCFG=true
A BLS configuration is the default and in this case the installkernel-bls
script is used. For a non-BLS configuration the grubby-deprecated package
installs an installkernel that uses the new-kernel-pkg script.

So there is no need to check for GRUB_ENABLE_BLSCFG=true. Also, this check
is wrong for s390x and was causing the installkernel-bls script to no call
kernel-install on this platform.

Resolves: rhbz#1647721

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>